Remove agent auth
# Conflicts:
#	components/director/pkg/graphql/schema.graphql
#	components/director/pkg/graphql/schema_gen.go
