+++ chart/compass/values.yaml
@@ -10,7 +10,7 @@ global:
       version: "PR-247"
     director:
       dir: pr/
-      version: "PR-319"
+      version: "PR-323"
     gateway:
       dir: pr/
       version: "PR-251"
@@ -19,7 +19,7 @@ global:
       version: "PR-99"
     schema_migrator:
       dir: pr/
-      version: "PR-319"
+      version: "PR-323"
     provisioner:
       dir: pr/
       version: "PR-305"
@@ -31,7 +31,7 @@ global:
     tests:
       e2e:
         dir: pr/
-        version: "PR-298"
+        version: "PR-323"
       connector:
         dir: pr/
         version: "PR-247"
