+++ components/director/internal/domain/runtime/converter_test.go
@@ -3,8 +3,6 @@ package runtime_test
 import (
 	"testing"
 
-	rtmautomock "github.com/kyma-incubator/compass/components/director/internal/domain/runtime/automock"
-
 	"github.com/kyma-incubator/compass/components/director/internal/domain/runtime"
 	"github.com/kyma-incubator/compass/components/director/internal/model"
 	"github.com/kyma-incubator/compass/components/director/pkg/graphql"
@@ -14,24 +12,17 @@ import (
 func TestConverter_ToGraphQL(t *testing.T) {
 	allDetailsInput := fixDetailedModelRuntime(t, "foo", "Foo", "Lorem ipsum")
 	allDetailsExpected := fixDetailedGQLRuntime(t, "foo", "Foo", "Lorem ipsum")
-	var modelAuth *model.Auth
 
 	// given
 	testCases := []struct {
-		Name            string
-		Input           *model.Runtime
-		Expected        *graphql.Runtime
-		AuthConverterFn func() *rtmautomock.AuthConverter
+		Name     string
+		Input    *model.Runtime
+		Expected *graphql.Runtime
 	}{
 		{
 			Name:     "All properties given",
 			Input:    allDetailsInput,
 			Expected: allDetailsExpected,
-			AuthConverterFn: func() *rtmautomock.AuthConverter {
-				conv := &rtmautomock.AuthConverter{}
-				conv.On("ToGraphQL", allDetailsInput.AgentAuth).Return(allDetailsExpected.AgentAuth).Once()
-				return conv
-			},
 		},
 		{
 			Name:  "Empty",
@@ -41,45 +32,28 @@ func TestConverter_ToGraphQL(t *testing.T) {
 					Condition: graphql.RuntimeStatusConditionInitial,
 				},
 			},
-			AuthConverterFn: func() *rtmautomock.AuthConverter {
-				conv := &rtmautomock.AuthConverter{}
-				conv.On("ToGraphQL", modelAuth).Return(nil).Once()
-				return conv
-			},
 		},
 		{
 			Name:     "Nil",
 			Input:    nil,
 			Expected: nil,
-			AuthConverterFn: func() *rtmautomock.AuthConverter {
-				conv := &rtmautomock.AuthConverter{}
-				return conv
-			},
 		},
 	}
 
 	for _, testCase := range testCases {
 		t.Run(testCase.Name, func(t *testing.T) {
-			authConverter := testCase.AuthConverterFn()
-
 			// when
-			converter := runtime.NewConverter(authConverter)
+			converter := runtime.NewConverter()
 			res := converter.ToGraphQL(testCase.Input)
 
 			// then
 			assert.Equal(t, testCase.Expected, res)
-
-			authConverter.AssertExpectations(t)
 		})
 	}
 }
 
 func TestConverter_MultipleToGraphQL(t *testing.T) {
 	// given
-	var modelAuth *model.Auth
-	authConverter := &rtmautomock.AuthConverter{}
-	authConverter.On("ToGraphQL", modelAuth).Return(nil)
-
 	input := []*model.Runtime{
 		fixModelRuntime("foo", "tenant-foo", "Foo", "Lorem ipsum"),
 		fixModelRuntime("bar", "tenant-bar", "Bar", "Dolor sit amet"),
@@ -97,7 +71,7 @@ func TestConverter_MultipleToGraphQL(t *testing.T) {
 	}
 
 	// when
-	converter := runtime.NewConverter(authConverter)
+	converter := runtime.NewConverter()
 	res := converter.MultipleToGraphQL(input)
 
 	// then
@@ -106,7 +80,6 @@ func TestConverter_MultipleToGraphQL(t *testing.T) {
 
 func TestConverter_InputFromGraphQL(t *testing.T) {
 	// given
-	authConverter := &rtmautomock.AuthConverter{}
 	testCases := []struct {
 		Name     string
 		Input    graphql.RuntimeInput
@@ -127,7 +100,7 @@ func TestConverter_InputFromGraphQL(t *testing.T) {
 	for _, testCase := range testCases {
 		t.Run(testCase.Name, func(t *testing.T) {
 			// when
-			converter := runtime.NewConverter(authConverter)
+			converter := runtime.NewConverter()
 			res := converter.InputFromGraphQL(testCase.Input)
 
 			// then
