+++ components/director/internal/domain/runtime/fixtures_test.go
@@ -70,27 +70,12 @@ func fixDetailedModelRuntime(t *testing.T, id, name, description string) *model.
 		Name:        name,
 		Description: &description,
 		Tenant:      "tenant",
-		AgentAuth: &model.Auth{
-			AdditionalHeaders: map[string][]string{
-				"test": {"bar"},
-			},
-			Credential: model.CredentialData{
-				Basic: &model.BasicCredentialData{
-					Username: "foo",
-					Password: "bar",
-				},
-			},
-		},
 	}
 }
 
 func fixDetailedGQLRuntime(t *testing.T, id, name, description string) *graphql.Runtime {
 	time, err := time.Parse(time.RFC3339, "2002-10-02T10:00:00-05:00")
 	require.NoError(t, err)
-
-	headers := graphql.HttpHeaders{
-		"test": {"bar"},
-	}
 
 	return &graphql.Runtime{
 		ID: id,
@@ -100,13 +85,6 @@ func fixDetailedGQLRuntime(t *testing.T, id, name, description string) *graphql.
 		},
 		Name:        name,
 		Description: &description,
-		AgentAuth: &graphql.Auth{
-			AdditionalHeaders: &headers,
-			Credential: graphql.BasicCredentialData{
-				Username: "foo",
-				Password: "bar",
-			},
-		},
 	}
 }
